✔ Returns the floor of the float

+++ 0x00-python_variable_annotations/2-floor.py
@@ -7,7 +7,8 @@
 Floor module
 """
 
-def floor(n: float) -> float:
+
+def floor(n: float) -> int:
     """
     Returns the floor of the float.
     """
